Move Main and Room into pages
