+++ src/components/App.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import GlobalStyles from "../styles/globalStyles";
-import Main from "./Main";
-import Room from "./Room";
+import Main from "./pages/Main";
+import Room from "./pages/Room";
 
 function App() {
     return (
